VisionCore: Move internal functions to own class for better testing.
